use _dt functions for reading/writing
